renames lib -> src, moves index in there
